skill added

+++ src/components/AdminComponents/EditSkillsBox/EditSkillsBox.tsx
@@ -1,7 +1,6 @@
 import React, { ChangeEvent, useEffect, useState } from "react";
 import styles from "./style.module.scss";
 import DefaultToogle from "../ToogleBtn/ToogleBtn";
-import StackTable from "../StackTable/StackTable";
 import axios from "axios";
 import Swal from "sweetalert2";
 import { useSession } from "next-auth/react";
@@ -10,13 +9,16 @@ import { convertToBase64 } from "@/utils/Function";
 import { nanoid } from "nanoid";
 import Loader from "../Loader/Loader";
 import InputBox from "../InputBox/InputBox";
-
+import TechSelectModal from "../TechBox/TechBox";
+import Data from "@/utils/IconList";
+import Image from "next/image";
+import LucideIcons from "@/icons/LucideIcons";
 type SkillType = {
-  _id: string;
-  name: string;
+  tech_name: string;
   image: string;
-  active: boolean;
-};
+  _id: string | number;
+  isSelected: boolean;
+}[];
 
 type props = {
   data:
@@ -30,37 +32,16 @@ type props = {
 };
 function EditSkillsBox(props: props) {
   const { data: session } = useSession();
-  const [skills, setSkills] = useState<SkillType[]>([]);
+  const [skills, setSkills] = useState<SkillType>([]);
   const [loading, setLoading] = useState<boolean>(true);
   const [submitLoading, setSubmitLoading] = useState<boolean>(false);
+  const [iconModal, setIconModal] = useState<boolean>(false);
 
-  const addNewTech = () => {
-    setSkills((prev) => [
-      ...prev,
-      {
-        name: "",
-        image: "",
-        _id: nanoid(5),
-        active: true,
-      },
-    ]);
-  };
-  const handleTechNameChange = (e: ChangeEvent<HTMLInputElement>, id: string) => {
+  const addNewTech = (id: string | number) => {
     setSkills((prev) => {
       return prev.map((item) => {
-        if (item._id == id) {
-          return { ...item, name: e.target.value };
-        } else {
-          return item;
-        }
-      });
-    });
-  };
-  const handleTechUrlChange = (e: ChangeEvent<HTMLTextAreaElement>, id: string) => {
-    setSkills((prev) => {
-      return prev.map((item) => {
-        if (item._id == id) {
-          return { ...item, image: e.target.value };
+        if (item._id === id) {
+          return { ...item, isSelected: true };
         } else {
           return item;
         }
@@ -68,38 +49,57 @@ function EditSkillsBox(props: props) {
     });
   };
 
-  const removeTechStack = (id: string) => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You want to remove this Skill?",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, delete it!",
-    }).then(async (result) => { 
-      if (result.isConfirmed) {
-        try {
-          const res = await axios.delete(`/api/skill/${id}`);
+  const removeTechStack = (id: string | number, name: string) => {
+    if (typeof id == "number") {
+      setSkills((prev) => {
+        return prev.map((item) => {
+          if (item._id === id) {
+            return { ...item, isSelected: false };
+          } else {
+            return item;
+          }
+        });
+      });
+    } else {
+      Swal.fire({
+        title: "Are you sure?",
+        text: `You want to remove ${name}?`,
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          try {
+            const res = await axios.delete(`/api/skill/${id}`);
 
-          if (res?.status == 200) {
-            const newArr = skills.filter((item) => item._id != id);
-            setSkills(newArr);
+            if (res?.status == 200) {
+              setSkills((prev) => {
+                return prev.map((item) => {
+                  if (item._id === id) {
+                    return { ...item, isSelected: false };
+                  } else {
+                    return item;
+                  }
+                });
+              });
+              Swal.fire({
+                title: "Deleted!",
+                text: "Skill has been deleted.",
+                icon: "success",
+              });
+            }
+          } catch (error: any) {
             Swal.fire({
-              title: "Deleted!",
-              text: "Skill has been deleted.",
-              icon: "success",
+              icon: "error",
+              title: "Oops...",
+              text: "Something went wrong!",
             });
           }
-        } catch (error: any) {
-          Swal.fire({
-            icon: "error",
-            title: "Oops...",
-            text: "Something went wrong!",
-          });
         }
-      }
-    });
+      });
+    }
   };
 
   const handleSubmit = async () => {
@@ -108,12 +108,12 @@ function EditSkillsBox(props: props) {
     }
     let techTempStack: any[] = [];
     skills.map((item) => {
-      if (item.image.length) {
+      if (item.isSelected) {
         techTempStack.push({
-          name: item.name,
+          name: item.tech_name,
           image: item.image,
           relation_id: session?.user.name,
-          active: item.active,
+          active: true,
           _id: item._id,
         });
       }
@@ -156,15 +156,23 @@ function EditSkillsBox(props: props) {
     try {
       const res = await axios.get("/api/skill");
       if (res?.status == 200) {
-        setSkills(res?.data?.data);
+        res?.data?.data.forEach((item: any) => {
+          setSkills((prevTechStack) =>
+            prevTechStack.map((tech) => (tech.tech_name === item.name ? { ...tech, _id: item._id, isSelected: true } : tech))
+          );
+        });
       }
     } catch (err: any) {
       console.log(err.message);
     }
     setLoading(false);
   };
+  const handleIconModal = () => {
+    setIconModal(!iconModal);
+  };
   useEffect(() => {
     FetchExistingSkills();
+    Data.map((item) => setSkills((prev) => [...prev, { ...item, isSelected: false }]));
   }, []);
   return (
     <div className={styles.container}>
@@ -180,39 +188,33 @@ function EditSkillsBox(props: props) {
         <div>
           <div className={styles.box}>
             <div className={styles.label}>Tech Stacks</div>
-            <div className={styles.add} onClick={addNewTech}>
+            <div className={styles.add} onClick={handleIconModal}>
               +
             </div>
             <div className={styles.techBox}>
-              {skills.map((tech) => (
-                <div key={tech._id} className={styles.techContainer}>
-                  <input
-                    type="text"
-                    name={tech._id}
-                    value={tech.name}
-                    onChange={(e) => handleTechNameChange(e, tech._id)}
-                    placeholder="Add Skill Name"
-                    className={styles.techInput}
-                  />
-                  <InputBox
-                    name="des"
-                    value={tech.image}
-                    handleTextAreaChange={(e) => handleTechUrlChange(e, tech._id)}
-                    handleChange={() => {}}
-                    type="textarea"
-                    placeholder="Paste Icon Url"
-                    row={3}
-                  />
-                  <div className={styles.techRebtn} onClick={() => removeTechStack(tech._id)}>
-                    Remove
+              {skills.map((tech: any) => {
+                return tech.isSelected ? (
+                  <div key={tech._id} className={tech.isSelected ? styles.tech_selected : styles.tech}>
+                    <Image src={tech.image} width={50} height={50} className={styles.tech_image} alt="" />
+                    <div className={styles.tech_name}>{tech.tech_name}</div>
+                    <div
+                      className={styles.tech_name}
+                      onClick={() => {
+                        removeTechStack(tech._id, tech.tech_name);
+                      }}
+                    >
+                      <LucideIcons name="delete" color="red" />
+                    </div>
                   </div>
-                </div>
-              ))}
+                ) : null;
+              })}
             </div>
           </div>
+
           <div className={styles.btn} onClick={handleSubmit}>
             {submitLoading ? "loading..." : "Submit"}
           </div>
+          <TechSelectModal open={iconModal} handleModal={handleIconModal} list={skills} onSelect={addNewTech} onDeSelect={() => {}} />
         </div>
       )}
     </div>
